requirements.txt
fastapi
uvicorn
pydantic
faiss-cpu
sentence-transformers
torch
transformers
scikit-learn
numpy
pickle-mixin
